requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
idna==3.10
numpy==2.2.4
psycopg==3.2.6
psycopg2-binary==2.9.10
pydub==0.25.1
PyJWT==2.1.0
python-core==0.0.1
requests==2.32.3
setuptools==78.1.0
SpeechRecognition==3.14.2
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
wheel==0.45.1
